Fix class/attribute mixup bug
There's a bug in the current version of GSP where an element like this:

    p .class attribute="value" {-}

would be transpiled as this:

    <p class="value" attribute="value"></p>

The `classes` slice in `func printAttrs` (`formatter.go`, line 50) was
the cause. I fixed it by making a deep copy of the `attrs` slice. I'm not
a Go programmer, there might be a nicer way to make a deep copy of a slice.
But it does, indeed, work.

P.S. thanks for making this, writing HTML by hand is now slightly less
painful!